Update blackJack.js
To start playing all you have to do is go on terminal/cmd and node the file; afterwards the game would start asking you prompts answer them; the goal of this game is to reach 21. It is getting a random card which you would use afterwards in the game that is within the array yourCards; In the end it adds up the values of yourCards bassed on the typical rules of blackjack; afterwards that answer determins whether you win or not.